requirements.txt
#python 3.6
easydict==1.13
kornia==0.6.8
lpips==0.1.4
matplotlib==3.3.4
numpy==1.19.5
opencv_python==3.4.3.18
Pillow==11.1.0
PyYAML==6.0.2
torch==1.10.1
torchvision==0.11.2
